chore: move {certificate,proposal,transaction}_pointer to their own modules.
